requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.17.0
pyinstaller==6.3.0
requests==2.31.0
psutil==5.9.8
Pillow==10.2.0 